DB-2600

+++ php_pdo_nuodb_c_cpp_common.h
@@ -53,9 +53,9 @@
 extern "C" {
 #endif
 struct pdo_nuodb_timer_t {
-    double startTimeInMicroSec;
-    double endTimeInMicroSec;
-    int    stopped;
+    int startTimeInMicroSec;
+    int endTimeInMicroSec;
+    int stopped;
 #ifdef WIN32
     LARGE_INTEGER frequency; // ticks per second
     LARGE_INTEGER startCount;
@@ -68,7 +68,7 @@ struct pdo_nuodb_timer_t {
 void pdo_nuodb_timer_init(struct pdo_nuodb_timer_t *timer);
 void pdo_nuodb_timer_start(struct pdo_nuodb_timer_t *timer);
 void pdo_nuodb_timer_end(struct pdo_nuodb_timer_t *timer);
-double pdo_nuodb_get_elapsed_time_in_microseconds(struct pdo_nuodb_timer_t *timer);
+int pdo_nuodb_get_elapsed_time_in_microseconds(struct pdo_nuodb_timer_t *timer);
 #ifdef __cplusplus
 }
 #endif
